GAN/requirements.txt
#Packages and Versions
torch == 2.3.1+cu121
tqdm == 4.66.4
numpy == 1.26.3
pillow == 10.2.0
matplotlib == 3.9.1
torchvision == 0.18.1+cu121
